+++ editor/src/components/menubar/menubar.tsx
@@ -15,8 +15,9 @@ import {
 import { betterReactMemo } from 'uuiui-deps'
 import { FLOATING_PREVIEW_BASE_URL } from '../../common/env-vars'
 import { LoginState } from '../../common/user'
-import { useReParseOpenProjectFile } from '../../core/model/project-file-test-utils'
+import { useReParseOpenProjectFile } from '../../core/model/project-file-helper-hooks'
 import { shareURLForProject } from '../../core/shared/utils'
+import { isFeatureEnabled } from '../../utils/feature-switches'
 import { EditorAction, EditorDispatch } from '../editor/action-types'
 import { setLeftMenuTab, setPanelVisibility, togglePanel } from '../editor/actions/actions'
 import { EditorState } from '../editor/store/editor-state'
@@ -183,9 +184,11 @@ export const Menubar = betterReactMemo('Menubar', () => {
           </span>
         </Tooltip>
       </FlexColumn>
-      <Tile style={{ marginTop: 12, marginBottom: 12 }}>
-        <a onClick={onReparseClick}>R</a>
-      </Tile>
+      {isFeatureEnabled('Re-parse Project Button') ? (
+        <Tile style={{ marginTop: 12, marginBottom: 12 }}>
+          <a onClick={onReparseClick}>R</a>
+        </Tile>
+      ) : null}
       <Tile style={{ marginTop: 12, marginBottom: 12 }}>
         <a href='/projects'>
           <Avatar loginState={userState.loginState} size={28} />
